pdf1_0,pdf1_7: change long long values to int values

+++ test/pdfio/pdf1_0/pdf1_0_indirect_object_testsuite.cpp
@@ -7,24 +7,24 @@ namespace pdf1_0 = pdfio::pdf1_0;
 TEST(IndirectObjectTestSuite, defaultCtor)
 {
 	pdf1_0::IndirectObject indirectObject;
-	EXPECT_TRUE(indirectObject.objectNumber() == 0ll);
-	EXPECT_TRUE(indirectObject.generationNumber() == 0ll);
+	EXPECT_TRUE(indirectObject.objectNumber() == 0);
+	EXPECT_TRUE(indirectObject.generationNumber() == 0);
 }
 
 TEST(IndirectObjectTestSuite, assignment)
 {
 	pdf1_0::IndirectObject indirectObject;
-	indirectObject.objectNumber() = 10ll;
-	indirectObject.generationNumber() = 1ll;
-	EXPECT_TRUE(indirectObject.objectNumber() == 10ll);
-	EXPECT_TRUE(indirectObject.generationNumber() == 1ll);
+	indirectObject.objectNumber() = 10;
+	indirectObject.generationNumber() = 1;
+	EXPECT_TRUE(indirectObject.objectNumber() == 10);
+	EXPECT_TRUE(indirectObject.generationNumber() == 1);
 }
 
 TEST(IndirectObjectTestSuite, setInteger)
 {
 	pdf1_0::IndirectObject indirectObject;
-	indirectObject.set(pdf1_0::Integer(23ll));
-	EXPECT_TRUE(indirectObject.get<pdf1_0::Integer>() == 23ll);
+	indirectObject.set(pdf1_0::Integer(23));
+	EXPECT_TRUE(indirectObject.get<pdf1_0::Integer>() == 23);
 	EXPECT_EQ(static_cast<int>(pdf1_0::GenericObjectType::kInteger), indirectObject.get().typeId_);
 }
 
@@ -39,9 +39,9 @@ TEST(IndirectObjectTestSuite, setName)
 TEST(IndirectObjectTestSuite, indirectReference)
 {
 	pdf1_0::IndirectObject indirectObject;
-	indirectObject.set(pdf1_0::IndirectReference(23ll, 2l));
-	EXPECT_TRUE(indirectObject.get<pdf1_0::IndirectReference>().objectNumber() == 23ll);
-	EXPECT_TRUE(indirectObject.get<pdf1_0::IndirectReference>().generationNumber() == 2ll);
+	indirectObject.set(pdf1_0::IndirectReference(23, 2));
+	EXPECT_TRUE(indirectObject.get<pdf1_0::IndirectReference>().objectNumber() == 23);
+	EXPECT_TRUE(indirectObject.get<pdf1_0::IndirectReference>().generationNumber() == 2);
 	EXPECT_EQ(static_cast<int>(pdf1_0::GenericObjectType::kIndirectReference), indirectObject.get().typeId_);
 }
 
@@ -49,10 +49,10 @@ TEST(IndirectObjectTestSuite, setDictionary)
 {
 	pdf1_0::IndirectObject indirectObject;
 	pdf1_0::Dictionary dictionary;
-	dictionary.insert(pdf1_0::Name("key1"), pdf1_0::Integer(56ll));
+	dictionary.insert(pdf1_0::Name("key1"), pdf1_0::Integer(56));
 	dictionary.insert(pdf1_0::Name("key2"), pdf1_0::Name("value2"));
 	indirectObject.set(dictionary);
-	EXPECT_TRUE(indirectObject.get<pdf1_0::Dictionary>().get<pdf1_0::Integer>(pdf1_0::Name("key1")) == 56ll);
+	EXPECT_TRUE(indirectObject.get<pdf1_0::Dictionary>().get<pdf1_0::Integer>(pdf1_0::Name("key1")) == 56);
 	EXPECT_TRUE(indirectObject.get<pdf1_0::Dictionary>().get<pdf1_0::Name>(pdf1_0::Name("key2")) == "value2");
 	EXPECT_EQ(static_cast<int>(pdf1_0::GenericObjectType::kDictionary), indirectObject.get().typeId_);
 }
